+++ resources/routes.php
@@ -1,4 +1,7 @@
 <?php
 
 Route::get(config('oauth2login.oauth_redirect_path'),
-    'Kronthto\LaravelOAuth2Login\OAuthLoginController@authorizeCallback')->middleware('web');
+    'Kronthto\LaravelOAuth2Login\OAuthLoginController@authorizeCallback')
+    ->middleware('web')
+    ->name('oauth2login.oauth_redirect_route')
+;